requirements.txt
asgiref==3.4.1
Django==3.2.5
django-crispy-forms==1.12.0
Pillow==8.3.1
pytz==2021.1
reportlab==3.5.68
sqlparse==0.4.1
sweetify==2.0.0
django-filter~=2.4.0